Added support for MotionChart
Add MotionChart and a corresponding MotionChartOptions class and
"motionchart" enum to the ChartPackage class.
Unlike the other charts MotionChart is not responsive. In order to keep the
same behavior as the other  chart types, the MotionChart will determine if
width or height are specified. If not it will assume to resize if the available
space. For this to work, the access level of the  Options variable inside of
the ChartWidget had to be changed from private to protected.

This fixes #53